Move sighash type tests to the sighash_type module
