+++ src/v2/map/input.rs
@@ -997,59 +997,7 @@ impl std::error::Error for CombineError {
 
 #[cfg(test)]
 mod test {
-    use core::str::FromStr;
-
     use super::*;
-
-    #[test]
-    fn psbt_sighash_type_ecdsa() {
-        for ecdsa in &[
-            EcdsaSighashType::All,
-            EcdsaSighashType::None,
-            EcdsaSighashType::Single,
-            EcdsaSighashType::AllPlusAnyoneCanPay,
-            EcdsaSighashType::NonePlusAnyoneCanPay,
-            EcdsaSighashType::SinglePlusAnyoneCanPay,
-        ] {
-            let sighash = PsbtSighashType::from(*ecdsa);
-            let s = format!("{}", sighash);
-            let back = PsbtSighashType::from_str(&s).unwrap();
-            assert_eq!(back, sighash);
-            assert_eq!(back.ecdsa_hash_ty().unwrap(), *ecdsa);
-        }
-    }
-
-    #[test]
-    fn psbt_sighash_type_taproot() {
-        for tap in &[
-            TapSighashType::Default,
-            TapSighashType::All,
-            TapSighashType::None,
-            TapSighashType::Single,
-            TapSighashType::AllPlusAnyoneCanPay,
-            TapSighashType::NonePlusAnyoneCanPay,
-            TapSighashType::SinglePlusAnyoneCanPay,
-        ] {
-            let sighash = PsbtSighashType::from(*tap);
-            let s = format!("{}", sighash);
-            let back = PsbtSighashType::from_str(&s).unwrap();
-            assert_eq!(back, sighash);
-            assert_eq!(back.taproot_hash_ty().unwrap(), *tap);
-        }
-    }
-
-    #[test]
-    fn psbt_sighash_type_notstd() {
-        let nonstd = 0xdddddddd;
-        let sighash = PsbtSighashType { inner: nonstd };
-        let s = format!("{}", sighash);
-        let back = PsbtSighashType::from_str(&s).unwrap();
-
-        assert_eq!(back, sighash);
-        // TODO: Uncomment this stuff.
-        // assert_eq!(back.ecdsa_hash_ty(), Err(NonStandardSighashTypeError(nonstd)));
-        // assert_eq!(back.taproot_hash_ty(), Err(InvalidSighashTypeError(nonstd)));
-    }
 
     #[cfg(feature = "std")]
     fn out_point() -> OutPoint {
